fix: query update and add response status and message
1. optimised query of groupFilter.
2. add response and handled data flow, when update, delete and get request for not present data.

+++ back-end/src/controller/group-controller.ts
@@ -11,16 +11,24 @@ export class GroupController {
   private groupStudentRepository = getRepository(GroupStudent)
 
   async allGroups(request: Request, response: Response, next: NextFunction) {
-    return this.groupRepository.find()
+    const groupsResponse = await this.groupRepository.find()
+    if (!groupsResponse) {
+      return { status: 400, message: "Missing/Invalid Data" }
+    }
+    return groupsResponse
   }
 
   async getGroup(request: Request, response: Response, next: NextFunction) {
-    return this.groupRepository.find(request.body.id)
+    const groupResponse = await this.groupRepository.findOne(request.body.id)
+    if (!groupResponse) {
+      return { status: 400, message: "Missing/Invalid Data" }
+    }
+    return groupResponse
   }
 
   async createGroup(request: Request, response: Response, next: NextFunction) {
     const { body: params } = request
-    console.log(params)
+
     const createGroupInput: CreateGroupInput = {
       name: params.name,
       number_of_weeks: params.number_of_weeks,
@@ -31,14 +39,20 @@ export class GroupController {
     const group = new Group()
     group.prepareToCreate(createGroupInput)
 
-    return this.groupRepository.save(group)
+    const groupResponse = await this.groupRepository.save(group)
+    if (!groupResponse) {
+      return { status: 400, message: "Missing/Invalid Data" }
+    }
+    return groupResponse
   }
 
   async updateGroup(request: Request, response: Response, next: NextFunction) {
     const { body: params } = request
 
-    // add logic to handle update request for undefined id or deleted id data
-    return this.groupRepository.findOne(params.id).then((group) => {
+    return this.groupRepository.findOne(params.id).then(async (group) => {
+      if (!group) {
+        return { status: 204, message: "No Content Found" }
+      }
       const updateGroupInput: UpdateGroupInput = {
         id: params.id,
         name: params.name,
@@ -49,18 +63,36 @@ export class GroupController {
       }
       group.prepareToUpdate(updateGroupInput)
 
-      return this.groupRepository.save(group)
+      const updateResponse = await this.groupRepository.save(group)
+      if (!updateResponse) {
+        return { status: 400, message: "Missing/Invalid Data" }
+      }
+      return updateResponse
     })
   }
 
   async removeGroup(request: Request, response: Response, next: NextFunction) {
     let groupToRemove = await this.groupRepository.findOne(request.body.id)
-    return await this.groupRepository.remove(groupToRemove)
+    if (!groupToRemove) {
+      return { status: 204, message: "No content Found" }
+    }
+    const removeResponse = await this.groupRepository.remove(groupToRemove)
+    if (!removeResponse) {
+      return { status: 400, message: "Missing/Invalid Data" }
+    }
+    return removeResponse
   }
 
   async getGroupStudents(request: Request, response: Response, next: NextFunction) {
-    let { number_of_weeks, ltmt, incidents, roll_states } = await this.groupRepository.findOne(request.body.id)
+    let findResponse = await this.groupRepository.findOne(request.body.id)
+    if (!findResponse) {
+      return { status: 204, message: "No data found" }
+    }
+    let { number_of_weeks, ltmt, incidents, roll_states } = findResponse
     let allGroupStudent = await this.groupStudentList(number_of_weeks, ltmt, incidents, roll_states)
+    if (!allGroupStudent) {
+      return { status: 204, message: "No data found" }
+    }
     allGroupStudent = allGroupStudent.map((groupStudent) => {
       const { student_id, first_name, last_name, full_name } = groupStudent
       return { student_id, first_name, last_name, full_name }
@@ -72,18 +104,16 @@ export class GroupController {
     const entityManager = getManager()
 
     return await entityManager.query(
-      `SELECT roll_id,name as roll_name,student_id, first_name, last_name,
-        (first_name || '  ' ||last_name) as full_name,state,incident_count
-      from (
-        SELECT  srs.roll_id, r.name,srs.student_id,s.first_name,s.last_name,srs.state, count(*) as incident_count FROM student_roll_state srs
+      `SELECT  srs.roll_id, r.name,srs.student_id,s.first_name,s.last_name,srs.state, count(*) as incident_count, 
+        (first_name || '  ' ||last_name) as full_name
+        FROM student_roll_state srs
         INNER JOIN student as s on s.id = srs.student_id
         INNER JOIN roll as r on r.id = srs.roll_id
         WHERE srs.created_at > date('now', '` +
         -7 * number_of_weeks +
         ` days')
         group by srs.state
-      ) as incident_data
-      WHERE incident_count ` +
+        having incident_count ` +
         ltmt +
         ` ` +
         incidents +
@@ -94,8 +124,7 @@ export class GroupController {
   }
 
   private async removeAllGroupStudent() {
-    let groupStudentToRemove = await this.groupStudentRepository.find()
-    return await this.groupStudentRepository.remove(groupStudentToRemove)
+    return await this.groupStudentRepository.clear()
   }
 
   private async addGroupStudent(groupStudentInput: CreateGroupStudentInput[], groupId: number) {
@@ -114,29 +143,37 @@ export class GroupController {
   }
 
   async runGroupFilters(request: Request, response: Response, next: NextFunction) {
-    const removeData = await this.removeAllGroupStudent()
+    await this.removeAllGroupStudent()
     let allGroup = await this.groupRepository.find()
+    if (!allGroup) {
+      return { status: 204, message: "No data found" }
+    }
     const groupStudent = map(allGroup, async (group) => {
       const groupStudents = await this.groupStudentList(group.number_of_weeks, group.ltmt, group.incidents, group.roll_states)
       if (groupStudents.length > 0) {
-        this.addGroupStudent(groupStudents, group.id)
+        const addResponse = await this.addGroupStudent(groupStudents, group.id)
+        if (!addResponse) {
+          return { status: 500, message: "Internal Server Error" }
+        }
       }
       const run_at = new Date()
-      this.groupRepository.findOne(group.id).then((group) => {
-        const updateGroupInput: UpdateGroupInput = {
-          id: group.id,
-          name: group.name,
-          number_of_weeks: group.number_of_weeks,
-          roll_states: group.roll_states,
-          incidents: group.incidents,
-          ltmt: group.ltmt,
-          run_at: run_at,
-          student_count: groupStudents.length,
-        }
-        group.prepareToUpdate(updateGroupInput)
+      const updateGroupInput: UpdateGroupInput = {
+        id: group.id,
+        name: group.name,
+        number_of_weeks: group.number_of_weeks,
+        roll_states: group.roll_states,
+        incidents: group.incidents,
+        ltmt: group.ltmt,
+        run_at: run_at,
+        student_count: groupStudents.length,
+      }
+      group.prepareToUpdate(updateGroupInput)
 
-        return this.groupRepository.save(group)
-      })
+      const updateResponse = await this.groupRepository.save(group)
+      if (!updateResponse) {
+        return { status: 400, message: "Missing/Invalid Data" }
+      }
+      return updateResponse
     })
     if (groupStudent.length > 0) {
       return true
